+++ projects/iad-interface-admin/src/public_api.ts
@@ -14,10 +14,11 @@ export * from './lib/iad-base-grid/model/iad-grid-column.model';
 export * from './lib/iad-base-grid/model/iad-grid-model';
 export * from './lib/iad-base-grid/model/column-components.model';
 export * from './lib/iad-base-grid/base-grid/base-grid-columns.service';
-export * from './lib/iad-base-grid/table-header-menu/table-header-menu.component';
 export * from './lib/iad-base-grid/base-grid/base-grid.component';
 export * from './lib/iad-base-grid/base-grid/base-grid-freeze-column.model';
+export * from './lib/iad-base-grid/table-header-menu/table-header-menu.component';
 export * from './lib/iad-base-grid/iad-base-grid.module';
+export * from './lib/iad-base-grid/services/iad-base-grid-actions-service';
 
 // # projection-tree
 export * from './lib/projection-tree/projection-tree.model';
@@ -26,7 +27,6 @@ export * from './lib/projection-tree/projection-tree.model';
 export * from './lib/projection-grid/model/projection-grid.model';
 export * from './lib/projection-grid/projection-grid.module';
 export * from './lib/projection-grid/projection-grid/projection-grid.component';
-export * from './lib/projection-grid/grid-toolbar/grid-toolbar.component';
 export * from './lib/projection-grid/settings-manager/grid-settings-storage.model';
 export * from './lib/projection-grid/settings-manager/grid-settings-storage.service';
 
@@ -34,9 +34,9 @@ export * from './lib/projection-grid/settings-manager/grid-settings-storage.serv
 export * from './lib/public-services/iad-projection-tree.service';
 
 // #table toolbar
-export * from './lib/toolbar/toolbar-action.directive';
-export * from './lib/toolbar/toolbar-actions.component';
-export * from './lib/toolbar/toolbar-action.model';
+export * from './lib/toolbar/iad-toolbar-action.directive';
+export * from './lib/toolbar/iad-toolbar-actions.component';
+export * from './lib/toolbar/iad-toolbar-action.model';
 export * from './lib/toolbar/iad-toolbar.module';
 
 // # main lib module:-